Rename: paginationSSR이름 변경
