+++ src/components/common/Pagination/Pagination.tsx
@@ -1,17 +1,15 @@
-'use client';
-
 import { ARROW_NEXT, ARROW_PREV } from '@/utils/constants';
-import { Dispatch, SetStateAction } from 'react';
+import Link from 'next/link';
 import styles from './Pagination.module.scss';
 
 type Props = {
   currentPage: number;
-  onPageChange: Dispatch<SetStateAction<number>>;
   allDataCount: number;
   perPageDataCount: number;
+  pageQuery?: string;
 };
 
-const PAGE_LIMIT = 7;
+const PAGE_LIMIT = 4;
 
 const createPageNumberList = (currentPage: number, totalPages: number): number[] => {
   const startPage = Math.max(1, Math.min(currentPage - 3, totalPages - PAGE_LIMIT + 1));
@@ -20,44 +18,46 @@ const createPageNumberList = (currentPage: number, totalPages: number): number[]
   return new Array(endPage - startPage + 1).fill(0).map((_, index) => index + startPage);
 };
 
-export default function Pagination({ currentPage, onPageChange, allDataCount, perPageDataCount }: Props) {
+export default function Pagination({ currentPage = 1, allDataCount, perPageDataCount, pageQuery }: Props) {
   const totalPages = Math.ceil(allDataCount / perPageDataCount);
   const pageNumberList = createPageNumberList(currentPage, totalPages);
 
   return (
     <div className={styles.container}>
       {totalPages > PAGE_LIMIT && (
-        <button
-          disabled={currentPage === 1}
-          type="button"
+        <Link
+          href={pageQuery ? `${pageQuery}&page=${currentPage - 1}` : `?page=${currentPage - 1}`}
           className={styles.button}
-          onClick={() => onPageChange((page) => page - 1)}
+          prefetch
         >
           <img src={ARROW_PREV} alt="arrow_prev" />
-        </button>
+        </Link>
       )}
       <ul className={styles.list}>
         {pageNumberList.map(
           (item) =>
             item <= totalPages && (
               <li key={item} className={`${styles.listItem} ${currentPage === item ? styles.selected : ''}`}>
-                <button className={styles.pageButton} type="button" onClick={() => onPageChange(item)}>
+                <Link
+                  href={pageQuery ? `${pageQuery}&page=${item}` : `?page=${item}`}
+                  className={styles.pageButton}
+                  prefetch
+                >
                   {item}
-                </button>
+                </Link>
               </li>
             ),
         )}
       </ul>
 
       {totalPages > PAGE_LIMIT && (
-        <button
-          disabled={currentPage === totalPages}
-          type="button"
+        <Link
+          href={pageQuery ? `${pageQuery}&page=${currentPage + 1}` : `?page=${currentPage + 1}`}
           className={styles.button}
-          onClick={() => onPageChange((page) => page + 1)}
+          prefetch
         >
           <img src={ARROW_NEXT} alt="arrow_prev" />
-        </button>
+        </Link>
       )}
     </div>
   );
