--- src/components/common/Pagination/index.ts
@@ -1,3 +1,0 @@
-import Pagination from './Pagination';
-
-export default Pagination;
