--- src/components/common/PaginationSSR/PaginationSSR.module.scss
@@ -1,58 +1,0 @@
-@import '@/styles/_index';
-
-.container {
-  display: flex;
-  align-items: center;
-  gap: 3rem;
-  justify-content: center;
-  background-color: #fff;
-  padding: 1rem 0;
-}
-
-.listItem {
-  width: 4rem;
-  height: 4rem;
-  align-items: center;
-  list-style-type: none;
-
-  cursor: pointer;
-  color: #111322;
-}
-
-.list {
-  display: flex;
-}
-
-.listItem.selected {
-  .pageButton {
-    color: $white;
-    background-color: $red-300;
-    border-radius: 0.4rem;
-  }
-}
-
-.pageButton {
-  background-color: $white;
-  border: none;
-  width: 100%;
-  height: 100%;
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.button {
-  background-color: #fff;
-  border: none;
-  cursor: pointer;
-
-  img {
-    vertical-align: top;
-  }
-}
-
-.link {
-  width: 100%;
-  height: 100%;
-}
